# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API Documentation",
        "description": "Blog API Documentation",
        "contact": {
            "email": "to.tahina@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://blog.test/api",
            "description": "Blog API Server"
        }
    ],
    "paths": {
        "/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Gets list of comments",
                "description": "Returns all comments",
                "operationId": "indexComment",
                "parameters": [
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Field to apply the sorting on",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "title"
                    },
                    {
                        "name": "orderDirection",
                        "in": "query",
                        "description": "Sorting direction DESC or ASC - Default ASC",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "asc"
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Numbe of records per page - Default 10",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/posts/{id}/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Lists comments on a single blog post",
                "description": "Lists comments on a single blog post",
                "operationId": "postComments",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Creates a comment",
                "description": "Creates a comment for a blog post",
                "operationId": "createPostComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreCommentRequest"
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/posts/{postId}/comments/{commentId}": {
            "put": {
                "tags": [
                    "Comments"
                ],
                "summary": "Updates a comment",
                "description": "Updates a comment for a blog post",
                "operationId": "updatePostComment",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreCommentRequest"
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
        "/comments/{id}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Deletes a comment",
                "description": "Deletes a comment",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Status of the operation",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "The response message",
                                            "type": "string",
                                            "example": "Comment successfully deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearer"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CommentResource": {
                "title": "CommentResource",
                "description": "Comment resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        }
                    },
                    "links": {
                        "title": "Links",
                        "description": "Pagination links",
                        "type": "array",
                        "format": "array",
                        "items": {
                            "properties": {
                                "first": {
                                    "type": "string",
                                    "example": "https://blog.test/api/posts?page=1"
                                },
                                "last": {
                                    "type": "string",
                                    "example": "https://blog.test/api/posts?page=3"
                                },
                                "prev": {
                                    "type": "string",
                                    "example": "https://blog.test/api/posts?page=3"
                                },
                                "next": {
                                    "type": "string",
                                    "example": "https://blog.test/api/posts?page=3"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "meta": {
                        "title": "Meta",
                        "description": "Meta data",
                        "type": "array",
                        "format": "array",
                        "items": {
                            "properties": {
                                "current_page": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "from": {
                                    "type": "integer",
                                    "example": 0
                                },
                                "last_page": {
                                    "type": "integer",
                                    "example": 5
                                },
                                "links": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "url": {
                                                "type": "string",
                                                "example": "http://blog.test/api/comments?page=1"
                                            },
                                            "label": {
                                                "type": "string",
                                                "example": "&laquo; Previous"
                                            },
                                            "active": {
                                                "type": "boolean",
                                                "example": true
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "path": {
                                    "type": "string",
                                    "example": "http://blog.test/api/comments"
                                },
                                "per_page": {
                                    "type": "integer",
                                    "example": 10
                                },
                                "to": {
                                    "type": "integer",
                                    "example": 10
                                },
                                "total": {
                                    "description": "Total number of comments",
                                    "type": "integer",
                                    "example": 100
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CommentResource"
                }
            },
...
            "Comment": {
                "title": "Comment",
                "description": "Comment model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "comment": {
                        "title": "Comment",
                        "description": "The blog post comment",
                        "type": "string",
                        "format": "string",
                        "example": "This is my first comment"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Creation date and time",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01 09:01:00"
                    },
                    "updated_at": {
                        "title": "Created at",
                        "description": "Creation date and time",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01 09:01:00"
                    },
                    "author": {
                        "$ref": "#/components/schemas/User"
                    },
                    "post": {
                        "$ref": "#/components/schemas/Post"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Comment"
                }
            },
            "Post": {
                "title": "Post",
                "description": "Post model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "The blog post title",
                        "type": "string",
                        "format": "string",
                        "example": "This is my first blog post"
                    },
                    "slug": {
                        "title": "Slug",
                        "description": "The blog post slug",
                        "type": "string",
                        "format": "string",
                        "example": "this-is-my-first-blog-post"
                    },
                    "body": {
                        "title": "Body",
                        "description": "The blog post content",
                        "type": "string",
                        "format": "string",
                        "example": "This is the content of my first blog"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Creation date and time",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01 09:01:00"
                    },
                    "updated_at": {
                        "title": "Created at",
                        "description": "Creation date and time",
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-01-01 09:01:00"
                    },
                    "author_id": {
                        "title": "Author ID",
                        "description": "The blog post author ID",
                        "type": "string",
                        "format": "integer",
                        "example": 22
                    },
                    "author": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Post"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "User's name",
                        "type": "string",
                        "format": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "title": "E-mail",
                        "description": "User's e-mail address",
                        "type": "string",
                        "format": "string",
                        "example": "john@doe.com"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "requestBodies": {
            "StoreCommentRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "comment"
                            ],
                            "properties": {
                                "comment": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "bearerFormat": "JWT"
            }
        }
    }
}